[helper] Allow to customize the H2Protocol used in the helper functions
By allowing to pass a custom H2Protocol, it will be easier for users of
the library to still use the helper and be able to provide a class that
inherits from H2Protocol. This way users of the library can for instance
override `H2Protocol.connection_lost`, call `super` to still get the
expected behavior and perform other actions, like reconnecting for
instance.